+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Run pipenv update, bringing pytest version 4.3.1, psycopg-binary 2.7.7, etc
 - sr.ht and TravisCI configuration to disable emojis and animation to keep logs clean
 
+### Changed
+- Use vanilla requests library instead of SolrClient
+
 ## [0.9.0] - 2019-01-22
 ### Updated
 - pytest version 4.0.0